requirements.txt
pathlib
numpy
pandas
opencv-python
imutils
pillow
tensorflow
sklearn
tqdm